+++ src/main/java/dev/otectus/sophisticatedtab/client/prefs/PreferencesStorage.java
@@ -24,7 +24,7 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.google.gson.JsonSyntaxException;
 import dev.otectus.sophisticatedtab.client.prefs.BackpackTabPreferences.ProfileEntry;
-import net.minecraftforge.fml.loading.FMLPaths;
+import net.neoforged.fml.loading.FMLPaths;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;